+++ Assets/Scripts/architecture_business/use_cases/player_spacecraft/motor/SpacecraftPlayerGetSpeedMotorUseCase.cs
@@ -3,11 +3,11 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public interface SpacecraftPlayerGetSpeedMotorUseCase {
-    int invoke();
+    float invoke();
 }
 
 public class SpacecraftPlayerGetSpeedMotorUseCaseImpl : SpacecraftPlayerGetSpeedMotorUseCase
 {
     private SpacecraftPlayerMotorRepository repo = new SpacecraftPlayerMotorRepositoryImpl();
-    public int invoke() => repo.speedMotor;
+    public float invoke() => repo.speedMotor;
 }
