+++ Assets/Scripts/game/player/motor/movement_available/BaseMotorMovementPlayer.cs
@@ -3,6 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public interface BaseMotorMovementPlayer {
-    void updateSpeedMotor(int speedMotor);
+    void updateSpeedMotor(float speedMotor);
     void move();  
 }
